+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 0.7.1 - 2020-12-18
+
+- Allow CXX environment variable to override g++ for musl build (#500)
+- Add user-defined checker for server side (#502)
+
 # 0.7.0 - 2020-11-02
 
 - Add blocking callback to `EnvBuilder` (#474)